+++ CRM/DataprocessorSearch/Form/Search.php
@@ -79,6 +79,25 @@ class CRM_DataprocessorSearch_Form_Search extends CRM_DataprocessorSearch_Form_A
   }
 
   /**
+   * Returns an array with hidden columns
+   *
+   * @return array
+   */
+  protected function getHiddenFields() {
+    $hiddenFields = array();
+    if (!$this->isIdFieldVisible()) {
+      $hiddenFields[] = $this->getIdFieldName();
+    }
+    if ($this->isSubmitted() && isset($this->dataProcessorOutput['configuration']['expose_hidden_fields']) && $this->dataProcessorOutput['configuration']['expose_hidden_fields']) {
+      $submittedHiddenFields = isset($this->_formValues['hidden_fields']) && is_array($this->_formValues['hidden_fields']) ? $this->_formValues['hidden_fields'] : array();
+      $hiddenFields = array_merge($hiddenFields, $submittedHiddenFields);
+    } elseif (isset($this->dataProcessorOutput['configuration']['hidden_fields']) && is_array($this->dataProcessorOutput['configuration']['hidden_fields'])) {
+      $hiddenFields = array_merge($hiddenFields, $this->dataProcessorOutput['configuration']['hidden_fields']);
+    }
+    return $hiddenFields;
+  }
+
+  /**
    * Return altered rows
    *
    * Save the ids into the queryParams value. So that when an action is done on the selected record
@@ -120,6 +139,7 @@ class CRM_DataprocessorSearch_Form_Search extends CRM_DataprocessorSearch_Form_A
   protected function buildCriteriaForm() {
     parent::buildCriteriaForm();
     $this->buildAggregateForm();
+    $this->buildHiddenFieldsForm();
   }
 
   /**
@@ -128,10 +148,49 @@ class CRM_DataprocessorSearch_Form_Search extends CRM_DataprocessorSearch_Form_A
    * @return false|String
    */
   protected function getAdditionalCriteriaTemplate() {
+    $return = [];
     if (isset($this->dataProcessorOutput['configuration']['expose_aggregate']) && $this->dataProcessorOutput['configuration']['expose_aggregate']) {
-      return "CRM/DataprocessorSearch/Form/Criteria/AggregateCriteria.tpl";
+      $return[] = "CRM/DataprocessorSearch/Form/Criteria/AggregateCriteria.tpl";
     }
-    return false;
+    if (isset($this->dataProcessorOutput['configuration']['expose_hidden_fields']) && $this->dataProcessorOutput['configuration']['expose_hidden_fields']) {
+      $return[] = "CRM/DataprocessorSearch/Form/Criteria/HiddenFieldsCriteria.tpl";
+    }
+    return $return;
+  }
+
+  /**
+   * Build the aggregate form
+   */
+  protected function buildHiddenFieldsForm() {
+    if (!isset($this->dataProcessorOutput['configuration']['expose_hidden_fields']) || !$this->dataProcessorOutput['configuration']['expose_hidden_fields']) {
+      return;
+    }
+    $size = $this->getCriteriaElementSize();
+
+    $sizeClass = 'huge';
+    $minWidth = 'min-width: 250px;';
+    if ($size =='compact') {
+      $sizeClass = 'medium';
+      $minWidth = '';
+    }
+
+    $fields = array();
+    $defaults = array();
+    foreach ($this->dataProcessorClass->getDataFlow()->getOutputFieldHandlers() as $outputFieldHandler) {
+      $fields[$outputFieldHandler->getOutputFieldSpecification()->alias] = $outputFieldHandler->getOutputFieldSpecification()->title;
+    }
+    if (isset($this->dataProcessorOutput['configuration']['hidden_fields']) && is_array($this->dataProcessorOutput['configuration']['hidden_fields'])) {
+      $defaults = $this->dataProcessorOutput['configuration']['hidden_fields'];
+    }
+
+    $this->add('select', "hidden_fields", '', $fields, false, [
+      'style' => $minWidth,
+      'class' => 'crm-select2 '.$sizeClass,
+      'multiple' => TRUE,
+      'placeholder' => E::ts('- Select -'),
+    ]);
+
+    $this->setDefaults(['hidden_fields' => $defaults]);
   }
 
 
@@ -180,6 +239,9 @@ class CRM_DataprocessorSearch_Form_Search extends CRM_DataprocessorSearch_Form_A
   protected function hasExposedFilters() {
     $return = parent::hasExposedFilters();
     if (!$return && isset($this->dataProcessorOutput['configuration']['expose_aggregate']) && $this->dataProcessorOutput['configuration']['expose_aggregate']) {
+      $return = true;
+    }
+    if (!$return && isset($this->dataProcessorOutput['configuration']['expose_hidden_fields']) && $this->dataProcessorOutput['configuration']['expose_hidden_fields']) {
       $return = true;
     }
     return $return;